Fix split-package between core and client modules
