+++ client/src/main/java/com/orientechnologies/orient/client/binary/OChannelBinarySynchClient.java
@@ -17,7 +17,7 @@
  *  * For more information: http://www.orientechnologies.com
  *
  */
-package com.orientechnologies.orient.enterprise.channel.binary;
+package com.orientechnologies.orient.client.binary;
 
 import com.orientechnologies.orient.core.config.OContextConfiguration;
